Phase A (16ddc772): brk addresses now propagate as full u64 through the
SYSCALL fast path so that sbrk continues to work after user space moves
into the upper canonical-low half.

  kernel/src/syscall/handlers/process.rs
    - new pub fn sys_sbrk_u64(increment: i64) -> u64
      operates on full u64 brk values from the start
    - existing sys_sbrk(i32) -> u32 now delegates to sys_sbrk_u64

  kernel/src/syscall/mod.rs
    - syscall_dispatch_64 special-cases SYS_SBRK and routes to
      sys_sbrk_u64 alongside SYS_MMAP/SYS_MUNMAP

  libs/libsyscall/src/lib.rs
    - sbrk()'s failure check changed from `u32::MAX as u64` to `u64::MAX`
  libs/stdlib/src/process.rs
    - sbrk() uses syscall1_u64() and treats `usize::MAX` as failure
  libs/stdlib/src/heap.rs
    - heap init failure check follows: `brk == usize::MAX as u64`

Composes cleanly with the in-flight syscall-handler u32->u64 widening
on main; Phase A only touches the SBRK path.

Boot verified in QEMU/KVM (virtio-vga) through compositor spawn.

+++ libs/libsyscall/src/lib.rs
@@ -344,9 +344,12 @@ pub fn exit(code: u32) -> ! {
 
 /// Extend the process heap by `increment` bytes.
 /// Returns previous break address, or `u64::MAX` on failure.
+///
+/// SYS_SBRK now uses the u64 ABI on the SYSCALL fast path: brk addresses
+/// can live anywhere in the canonical-low half.
 pub fn sbrk(increment: u32) -> u64 {
     let ret = syscall1(SYS_SBRK, increment as u64);
-    if ret == u32::MAX as u64 { u64::MAX } else { ret }
+    if ret == u64::MAX { u64::MAX } else { ret }
 }
 
 /// Map anonymous pages. Returns address or `u64::MAX` on failure.
